scuzz: move to pkg/mount/scuzz
