+++ pkg/mount/scuzz/const_amd64.go
